+++ apps/cli/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix: Pass `flutter_assets` to backend when using `celest deploy`
 - fix: Sentry integration
+- fix: Ensure pub cache is fixed during startup
 
 ## 1.0.10+2